Added ( horribly ) a M2 REST elseif
I had to take a few comment out lines and tweak a  few things as I really just wanted REST to work. Lines 220-272. It should give you at least an idea on how to get the rest calls working. When I get some of that 'free' time, I'd love to do a rewrite of this tool with you.